Give correct instructions for Windows
It needs to be much more clear that X11 is for Linux and GDI is for Windows.  I don't know what Mac or other platforms would use here

Also the better way to compile with the Text IDE is to load the appropriate fpgui_toolkit.pas file
which will compile ALL units, not just what may be used in a particular project, then all the newly
created units can be used for various projects, and you won't run into a problem if your original
project didn't use one of the units.

The instructions for compiling with the IDE have been modified to show usage for Linux and Windows
and further explain how to compile fpgui_toolkit.pas to compile all units